pyproject.toml
[tool.poetry]
name = "cirro"
version = "0.8.3"
description = "CLI tool and SDK for interacting with the Cirro platform"
authors = ["Cirro Bio <support@cirro.bio>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/CirroBio/Cirro-client"
packages = [{include = "cirro"}]

[tool.poetry.dependencies]
python = "^3.9"
attrs = ">=21.3.0"
cirro_api_client = "0.0.5"
click = "^8.1.3"
boto3 = "1.33.13"
questionary = "^2.0.1"
tqdm = "^4.62.3"
pygithub = "2.1.1"
jsonschema = "^4.21.1"
pandas = "^2.2.0"
s3fs = "2023.12.2"
pyjwt = "^2.7.0"
msal-extensions = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-cov = "*"
flake8 = "*"

[tool.poetry.scripts]
cirro-cli = "cirro.cli.cli:main"
cirro = "cirro.cli.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
